# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Mutual Fund - Fund Management Service
  description: APIs helps to manage the fund details.
  version: 1.0.0

servers:
  - url: http://localhost:8081/fund-management/v1/api

paths:
  /v1/api/funds:
    post:
      tags:
        - FundController
      summary: Add new funds
      operationId: createFund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundView"
      responses:
        '201':
          description: Funds creation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundResponse"
    put:
      tags:
        - FundController
      summary: Update funds nav
      operationId: updateFund
      parameters:
        - name: fundId
          in: path
          required: true
          description: FundId to update fund nav and nav date
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundNavView"
      responses:
        '201':
          description: Funds creation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundResponse"

  /v1/api/funds/order:
    post:
      tags:
        - OrderController
      summary: Buy Order request
      operationId: createOrder
      parameters:
        - name: orderType
          in: query
          required: true
          schema:
            type: string
            enum: [ BUY, SELL ]
          description: Order type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderView"
      responses:
        '201':
          description: Order creation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"

components:
  schemas:
    FundResponse:
      type: object
      properties:
        fund:
          $ref: "#/components/schemas/FundView"
        code:
          type: integer
        message:
          type: string
      required:
        - status
        - message
        - fund
    FundView:
      type: object
      properties:
        fundId:
          type: string
          minLength: 10
          maxLength: 20
          pattern: "^[0-9]+$"
          description: "Must be numeric, 10-20 characters"
        fundName:
          type: string
          minLength: 10
          maxLength: 50
          pattern: "^[a-zA-Z0-9_ ]*$"
          description: "Must be alphanumeric, 10-50 characters"
        nav:
          type: number
          description: Script price
          minimum: 0.1
          example: 999.99
        totalUnits:
          type: number
          description: Total number of Units available
        navDate:
          type: string
          format: date
          description: Nav date
          example: 2025-05-31
      required:
        - fundId
        - fundName
        - nav
        - navDate
    FundNavView:
      type: object
      properties:
        nav:
          type: number
          format: double
          description: Script price
          minimum: 0.1
          example: 999.99
        navDate:
          type: string
          format: date
          description: Nav date
          example: 2025-05-31
      required:
        - nav
        - navDate
    OrderView:
      type: object
      properties:
        username:
          type: string
          minLength: 5
          maxLength: 20
          pattern: "^[a-zA-Z0-9_]+$"
          description: "Must be alphanumeric with underscores, 5-20 characters"
        fundId:
          type: string
          minLength: 10
          maxLength: 20
          pattern: "^[0-9]+$"
          description: "Must be numeric, 10-20 characters"
        units:
          type: number
          description: Number of units to buy
          minimum: 1
        nav:
          type: number
          description: Current Date Script price
          minimum: 0.1
          example: 999.99

    OrderResponse:
      type: object
      required:
        - totalValue
        - status
        - message
      properties:
        totalValue:
          type: number
        code:
          type: integer
        message:
          type: string